requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.0
asyncpg==0.30.0
bcrypt==4.2.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
passlib==1.7.4
pydantic==2.9.2
pydantic_core==2.23.4
pydentic==0.0.1.dev3
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.17
python-stdnum==1.20
PyYAML==6.0.2
redis==5.2.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
